requirements.txt
fastapi
uvicorn[standard]
pydantic
nltk
python-dotenv
groq
